Fix Regexp
Changing the regexp for S,F an C tag so that the last parameter is a string.
on S F tags (and even C tag) the last parameter is indeed a string even
if the cookie or counter are numbers.
using a %d instead of %s prevent to use a string cookie in F S tag
remove str conversion on event.value now useless

Signed-off-by: Stephane Gasparini <stephane.gasparini@intel.com>